requirements.txt
openai
python-dotenv
pydantic
langchain
bs4
tiktoken
streamlit
requests
Pillow
langchain-community
uvicorn[standard]
fastapi[all]